requirements.txt
akshare>=1.10.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
tushare>=1.3.0
efinance>=0.5.0
